+++ src/oss4energy/src/database/__init__.py
@@ -7,7 +7,7 @@ from datetime import UTC, datetime
 
 from sqlmodel import Field, Session, SQLModel, create_engine, select
 
-from oss4energy.config import SETTINGS
+from oss4energy.src.config import SETTINGS
 
 
 # -------------------------------------------------------------------------------------
